reduce indirection
