+++ packages/@expo/cli/src/export/fork-bundleAsync.ts
@@ -1,4 +1,4 @@
-import { ExpoConfig, getConfigFilePaths, Platform } from '@expo/config';
+import { ExpoConfig, getConfigFilePaths, Platform, ProjectConfig } from '@expo/config';
 import type { LoadOptions } from '@expo/metro-config';
 import chalk from 'chalk';
 import Metro, { AssetData } from 'metro';
@@ -18,6 +18,7 @@ import {
   importMetroFromProject,
   importMetroServerFromProject,
 } from '../start/server/metro/resolveFromProject';
+import { getEntryWithServerRoot } from '../start/server/middleware/ManifestMiddleware';
 
 export type MetroDevServerOptions = LoadOptions & {
   quiet?: boolean;
@@ -58,6 +59,51 @@ async function assertEngineMismatchAsync(
     configFilePath,
     platform,
     isHermesManaged
+  );
+}
+
+export async function createBundlesAsync(
+  projectRoot: string,
+  projectConfig: ProjectConfig,
+  bundleOptions: {
+    clear?: boolean;
+    maxWorkers?: number;
+    platforms: Platform[];
+    dev?: boolean;
+    minify?: boolean;
+    sourcemaps?: boolean;
+  }
+): Promise<Partial<Record<Platform, BundleOutput>>> {
+  if (!bundleOptions.platforms.length) {
+    return {};
+  }
+  const { exp } = projectConfig;
+
+  const bundles = await bundleAsync(
+    projectRoot,
+    exp,
+    {
+      // If not legacy, ignore the target option to prevent warnings from being thrown.
+      resetCache: bundleOptions.clear,
+      maxWorkers: bundleOptions.maxWorkers,
+      quiet: false,
+    },
+    bundleOptions.platforms.map((platform: Platform) => ({
+      sourcemaps: bundleOptions.sourcemaps,
+      platform,
+      entryPoint: getEntryWithServerRoot(projectRoot, projectConfig, platform),
+      minify: bundleOptions.minify,
+      dev: bundleOptions.dev,
+    }))
+  );
+
+  // { ios: bundle, android: bundle }
+  return bundleOptions.platforms.reduce<Partial<Record<Platform, BundleOutput>>>(
+    (prev, platform, index) => ({
+      ...prev,
+      [platform]: bundles[index],
+    }),
+    {}
   );
 }
